Select noone when nightaction is not required
